Move all UI-related code to a subdirectory.
