+++ src/ui/model.rs
@@ -13,7 +13,7 @@ use anyhow::Error;
 
 use crate::capture::CaptureReader;
 use crate::item::{TrafficItem, TrafficViewMode, DeviceItem, DeviceViewMode};
-use crate::tree_list_model::{TreeListModel, ItemNodeRc};
+use crate::ui::tree_list_model::{TreeListModel, ItemNodeRc};
 
 /// Trait implemented by each of our ListModel implementations.
 pub trait GenericModel<Item, ViewMode> where Self: Sized {
@@ -126,8 +126,8 @@ mod imp {
 
     use std::cell::RefCell;
     use crate::item::{TrafficItem, TrafficViewMode, DeviceItem, DeviceViewMode};
-    use crate::row_data::{TrafficRowData, DeviceRowData};
-    use crate::tree_list_model::TreeListModel;
+    use crate::ui::row_data::{TrafficRowData, DeviceRowData};
+    use crate::ui::tree_list_model::TreeListModel;
 
     /// Define the inner type to be used in the GObject type system.
     macro_rules! model {
